ANDR-5: Правки.Вторая функция в маппере приватная; Классы для тестов перенесены внутрь класса самих тестов; Сам тест сокращен до 1 функции и 1 аргументПровайдера
