+++ feature/interview-trainer/impl/src/main/java/ru/yeahub/interview_trainer/impl/createQuiz/data/CreateQuizDataToDomainMapper.kt
@@ -16,8 +16,7 @@ class CreateQuizDataToDomainMapper {
             total = dataResponse.total
         )
 
-    //функция публичная для тестов
-    fun dataToDomain(data: GetSpecializationResponse): DomainSpecialization =
+    private fun dataToDomain(data: GetSpecializationResponse): DomainSpecialization =
         DomainSpecialization(
             id = data.id,
             title = data.title
