+++ feature/interview-trainer/impl/src/test/java/test/CreateQuizMapperTest.kt
@@ -1,7 +1,6 @@
 package test
 
 import org.junit.jupiter.api.Assertions
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsSource
 import ru.yeahub.interview_trainer.impl.createQuiz.data.CreateQuizDataToDomainMapper
@@ -10,93 +9,76 @@ import ru.yeahub.interview_trainer.impl.createQuiz.domain.DomainSpecializationLi
 import ru.yeahub.network_api.models.GetSpecializationResponse
 import ru.yeahub.network_api.models.GetSpecializationsResponse
 import ru.yeahub.test.TestArgumentsProvider
-import test.SpecializationExampleDataClasses.defaultDomainSpecial
-import test.SpecializationExampleDataClasses.defaultDomainSpecialListResponse
-import test.SpecializationExampleDataClasses.defaultDomainSpecialWithImage
-import test.SpecializationExampleDataClasses.defaultSpecialListResponse
-import test.SpecializationExampleDataClasses.defaultSpecialResponse
-import test.SpecializationExampleDataClasses.defaultSpecialResponseWithImage
 
 class CreateQuizMapperTest {
     private val toDomainMapper = CreateQuizDataToDomainMapper()
 
-    //CreateQuizDataToDomainMapper тесты
-
-    @Test
-    fun dataListToDomainList_shouldHandleEmptyList() {
-        val emptyData = GetSpecializationsResponse(
-            page = 1L,
-            limit = 10L,
-            data = emptyList(),
-            total = 0L
-        )
-        val result = toDomainMapper.dataListToDomainList(emptyData)
-        val expected = DomainSpecializationListResponse(
-            page = 1L,
-            limit = 10L,
-            data = emptyList(),
-            total = 0L
-        )
-        Assertions.assertEquals(expected, result)
-    }
-
-    @Test
-    fun dataToDomain_shouldMapCorrectly() {
-        val result = toDomainMapper.dataToDomain(defaultSpecialResponse)
-        Assertions.assertEquals(defaultDomainSpecial, result)
-    }
-
-    //Параметризированные тесты
-
-    class ArgumentsProvider1 :
-        TestArgumentsProvider<SpecializationSelectionDataToDomainMapperTestCase1>() {
-        override fun testCases(): List<SpecializationSelectionDataToDomainMapperTestCase1> = listOf(
-            SpecializationSelectionDataToDomainMapperTestCase1(
-                dataToTest = defaultSpecialResponse,
-                expectedResult = defaultDomainSpecial
-            ),
-            SpecializationSelectionDataToDomainMapperTestCase1(
-                dataToTest = defaultSpecialResponseWithImage,
-                expectedResult = defaultDomainSpecialWithImage
-            )
-        )
-    }
-
-    class ArgumentsProvider2 :
-        TestArgumentsProvider<SpecializationSelectionDataToDomainMapperTestCase2>() {
-        override fun testCases(): List<SpecializationSelectionDataToDomainMapperTestCase2> = listOf(
-            SpecializationSelectionDataToDomainMapperTestCase2(
-                dataToTest = defaultSpecialListResponse,
-                expectedResult = defaultDomainSpecialListResponse
+    //CreateQuizDataToDomainMapper параметризированный тест
+    class ArgumentsProvider :
+        TestArgumentsProvider<SpecializationSelectionDataToDomainMapperTestCase>() {
+        override fun testCases(): List<SpecializationSelectionDataToDomainMapperTestCase> = listOf(
+            SpecializationSelectionDataToDomainMapperTestCase(
+                dataToTest = SpecializationExampleDataClasses.defaultSpecialListResponse,
+                expectedResult = SpecializationExampleDataClasses.defaultDomainSpecialListResponse
             )
         )
     }
 
     @ParameterizedTest
-    @ArgumentsSource(ArgumentsProvider1::class)
-    fun specializationSelectionDataToDomainMapperTestCase1(
-        testCase1: SpecializationSelectionDataToDomainMapperTestCase1,
+    @ArgumentsSource(ArgumentsProvider::class)
+    fun specializationSelectionDataToDomainMapperTestCase(
+        testCase: SpecializationSelectionDataToDomainMapperTestCase,
     ) {
-        val result = toDomainMapper.dataToDomain(testCase1.dataToTest)
-        Assertions.assertEquals(testCase1.expectedResult, result)
+        val result = toDomainMapper.dataListToDomainList(testCase.dataToTest)
+        Assertions.assertEquals(testCase.expectedResult, result)
     }
 
-    @ParameterizedTest
-    @ArgumentsSource(ArgumentsProvider2::class)
-    fun specializationSelectionDataToDomainMapperTestCase2(
-        testCase2: SpecializationSelectionDataToDomainMapperTestCase2,
-    ) {
-        val result = toDomainMapper.dataListToDomainList(testCase2.dataToTest)
-        Assertions.assertEquals(testCase2.expectedResult, result)
+    object SpecializationExampleDataClasses {
+        val defaultSpecialResponse = GetSpecializationResponse(
+            id = 0L,
+            title = "default title num 0",
+            description = "default description",
+            imageSrc = null,
+            createdAt = "01.01.1970",
+            updatedAt = "01.01.1970"
+        )
+
+        val defaultSpecialResponseWithImage = GetSpecializationResponse(
+            id = 1L,
+            title = "default title num 1",
+            description = "default description",
+            imageSrc = "some image url",
+            createdAt = "01.01.1970",
+            updatedAt = "01.01.1970"
+        )
+
+        val defaultDomainSpecial = DomainSpecialization(
+            id = 0L,
+            title = "default title num 0"
+        )
+
+        val defaultDomainSpecialWithImage = DomainSpecialization(
+            id = 1L,
+            title = "default title num 1"
+        )
+
+        val defaultSpecialListResponse = GetSpecializationsResponse(
+            page = 1L,
+            limit = 10L,
+            data = listOf(defaultSpecialResponse, defaultSpecialResponseWithImage),
+            total = 2L
+        )
+
+        val defaultDomainSpecialListResponse = DomainSpecializationListResponse(
+            page = 1L,
+            limit = 10L,
+            data = listOf(defaultDomainSpecial, defaultDomainSpecialWithImage),
+            total = 2L
+        )
     }
+
+    data class SpecializationSelectionDataToDomainMapperTestCase(
+        val dataToTest: GetSpecializationsResponse,
+        val expectedResult: DomainSpecializationListResponse,
+    )
 }
-
-data class SpecializationSelectionDataToDomainMapperTestCase1(
-    val dataToTest: GetSpecializationResponse,
-    val expectedResult: DomainSpecialization,
-)
-
-data class SpecializationSelectionDataToDomainMapperTestCase2(
-    val dataToTest: GetSpecializationsResponse,
-    val expectedResult: DomainSpecializationListResponse,
-)
